trade-opportunities-api/app/utils/auth.py
"""
JWT authentication and password hashing utilities
"""
import hashlib
import secrets
import uuid
from datetime import datetime, timedelta
from typing import Optional

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer

from app.config import get_settings
from app.utils.storage import storage

settings = get_settings()
bearer_scheme = HTTPBearer()


def hash_password(password: str) -> str:
    """SHA-256 hash with salt"""
    salt = "trade_opportunities_salt_2024"
    return hashlib.sha256(f"{salt}{password}".encode()).hexdigest()


def verify_password(plain: str, hashed: str) -> bool:
    return hash_password(plain) == hashed


def create_access_token(username: str) -> tuple[str, datetime]:
    expires = datetime.utcnow() + timedelta(minutes=settings.jwt_expire_minutes)
    payload = {
        "sub": username,
        "exp": expires,
        "iat": datetime.utcnow(),
        "jti": str(uuid.uuid4())
    }
    token = jwt.encode(payload, settings.jwt_secret_key, algorithm=settings.jwt_algorithm)
    storage.store_token(token, username)
    return token, expires


def decode_token(token: str) -> Optional[dict]:
    try:
        payload = jwt.decode(token, settings.jwt_secret_key, algorithms=[settings.jwt_algorithm])
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token has expired. Please login again.",
            headers={"WWW-Authenticate": "Bearer"}
        )
    except jwt.InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication token.",
            headers={"WWW-Authenticate": "Bearer"}
        )


def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme)) -> str:
    """Dependency: validates JWT and returns username"""
    token = credentials.credentials
    payload = decode_token(token)
    username = payload.get("sub")

    if not username:
        raise HTTPException(status_code=401, detail="Invalid token payload")

    user = storage.get_user(username)
    if not user:
        raise HTTPException(status_code=401, detail="User not found")

    return username


def generate_session_id() -> str:
    return str(uuid.uuid4())