trade-opportunities-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pyjwt==2.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12